refactor: listing component improvements
